+++ hvad/models.py
@@ -21,20 +21,21 @@ import warnings
 def update_settings(*args, **kwargs):
     global FALLBACK_LANGUAGES, TABLE_NAME_SEPARATOR
     FALLBACK_LANGUAGES = tuple( code for code, name in settings.LANGUAGES )
-    try:
-        TABLE_NAME_SEPARATOR = getattr(settings, 'NANI_TABLE_NAME_SEPARATOR')
-    except AttributeError:
-        TABLE_NAME_SEPARATOR = getattr(settings, 'HVAD_TABLE_NAME_SEPARATOR', '_')
-    else:
-        warnings.warn('NANI_TABLE_NAME_SEPARATOR setting is deprecated and will '
-                      'be removed. Please rename it to HVAD_TABLE_NAME_SEPARATOR.',
-                      DeprecationWarning)
+    TABLE_NAME_SEPARATOR = getattr(settings, 'HVAD_TABLE_NAME_SEPARATOR', '_')
+
+    if hasattr(settings, 'NANI_TABLE_NAME_SEPARATOR'):
+        # remove in 1.3
+        raise ImproperlyConfigured(
+            'NANI_TABLE_NAME_SEPARATOR setting is obsolete and has been '
+            'removed. Please rename it to HVAD_TABLE_NAME_SEPARATOR.')
+
 
 #===============================================================================
 
 def _split_together(constraints, fields, meta, name):
     sconst, tconst = [], []
     if name in meta:
+        # raise in 1.3, remove in 1.5
         warnings.warn('Passing \'%s\' to TranslatedFields is deprecated. Please use '
                       'Please Meta.%s instead.' % (name, name), DeprecationWarning)
         tconst.extend(meta[name])
@@ -175,12 +176,12 @@ class BaseTranslationModel(models.Model):
 
 class TranslatableModelBase(ModelBase):
     def __new__(cls, *args, **kwargs):
-        warnings.warn('TranslatableModelBase metaclass is deprecated and will '
-            'be removed. Hvad no longer uses a custom metaclass so conflict '
+        # remove in 1.3
+        raise RuntimeError(
+            'TranslatableModelBase metaclass is no longer used and has been '
+            'removed. Hvad no longer uses a custom metaclass so conflict '
             'resolution is no longer required, TranslatableModelBase can be '
-            'dropped.',
-            DeprecationWarning)
-        return ModelBase.__new__(cls, *args, **kwargs)
+            'dropped.')
 
 
 class NoTranslation(object):